requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain
langchain-core
langchain-groq
python-dotenv
tzlocal
requests
streamlit